refactor: move edit UI into separate component, remove useEffect
